Move fake class into the right package.
