+++ vector/src/test/java/im/vector/app/features/poll/create/CreatePollViewModelTest.kt
@@ -18,20 +18,20 @@ package im.vector.app.features.poll.create
 
 import com.airbnb.mvrx.test.MvRxTestRule
 import im.vector.app.features.poll.PollMode
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.A_FAKE_OPTIONS
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.A_FAKE_QUESTION
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.A_FAKE_ROOM_ID
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.A_POLL_START_TIMELINE_EVENT
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.createPollArgs
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.editPollArgs
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.editedPollViewState
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.initialCreatePollViewState
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.pollViewStateWithOnlyQuestion
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.pollViewStateWithQuestionAndEnoughOptions
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.pollViewStateWithQuestionAndEnoughOptionsButDeletedLastOption
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.pollViewStateWithQuestionAndMaxOptions
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.pollViewStateWithQuestionAndNotEnoughOptions
-import im.vector.app.features.poll.create.FakeCreatePollViewStates.pollViewStateWithoutQuestionAndEnoughOptions
+import im.vector.app.test.fakes.FakeCreatePollViewStates.A_FAKE_OPTIONS
+import im.vector.app.test.fakes.FakeCreatePollViewStates.A_FAKE_QUESTION
+import im.vector.app.test.fakes.FakeCreatePollViewStates.A_FAKE_ROOM_ID
+import im.vector.app.test.fakes.FakeCreatePollViewStates.A_POLL_START_TIMELINE_EVENT
+import im.vector.app.test.fakes.FakeCreatePollViewStates.createPollArgs
+import im.vector.app.test.fakes.FakeCreatePollViewStates.editPollArgs
+import im.vector.app.test.fakes.FakeCreatePollViewStates.editedPollViewState
+import im.vector.app.test.fakes.FakeCreatePollViewStates.initialCreatePollViewState
+import im.vector.app.test.fakes.FakeCreatePollViewStates.pollViewStateWithOnlyQuestion
+import im.vector.app.test.fakes.FakeCreatePollViewStates.pollViewStateWithQuestionAndEnoughOptions
+import im.vector.app.test.fakes.FakeCreatePollViewStates.pollViewStateWithQuestionAndEnoughOptionsButDeletedLastOption
+import im.vector.app.test.fakes.FakeCreatePollViewStates.pollViewStateWithQuestionAndMaxOptions
+import im.vector.app.test.fakes.FakeCreatePollViewStates.pollViewStateWithQuestionAndNotEnoughOptions
+import im.vector.app.test.fakes.FakeCreatePollViewStates.pollViewStateWithoutQuestionAndEnoughOptions
 import im.vector.app.test.fakes.FakeSession
 import im.vector.app.test.test
 import io.mockk.unmockkAll
