+++ vector/src/test/java/im/vector/app/test/fakes/FakeCreatePollViewStates.kt
@@ -14,9 +14,12 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.poll.create
+package im.vector.app.test.fakes
 
 import im.vector.app.features.poll.PollMode
+import im.vector.app.features.poll.create.CreatePollArgs
+import im.vector.app.features.poll.create.CreatePollViewModel
+import im.vector.app.features.poll.create.CreatePollViewState
 import org.matrix.android.sdk.api.session.events.model.Event
 import org.matrix.android.sdk.api.session.events.model.EventType
 import org.matrix.android.sdk.api.session.events.model.toContent
